backend/requirements.txt
# PyTorch and related libraries
torch==2.0.1
torchvision==0.15.2
transformers==4.31.0

# Pretrained model dependencies
timm==0.9.2  # Required for ViT model
bioc==2.1  # Required for Bio_ClinicalBERT

# Data Science and Utility Libraries
scikit-learn==1.3.0
matplotlib==3.7.1
numpy==1.24.3
scipy==1.11.1
tqdm==4.66.1
pyyaml==6.0

# FastAPI and related libraries
fastapi==0.95.1
pydantic==1.10.7
uvicorn==0.23.0

# Testing Framework
pytest==7.3.2
